docs: display correct default value for primitives instead of an empty string
.Also ignore CongitRoot and ConfigGroup static fields
