+++ core/processor/src/main/java/io/quarkus/annotation/processor/ExtensionAnnotationProcessor.java
@@ -685,7 +685,75 @@ public class ExtensionAnnotationProcessor extends AbstractProcessor {
             processingEnv.getMessager().printMessage(Diagnostic.Kind.ERROR, "Unable to find javadoc for config item " + e, e);
             return "";
         }
-        return docComment.trim();
+
+        return formatDocs(docComment.trim());
+    }
+
+    /**
+     * TODO - properly parse JavaDoc
+     */
+    private String formatDocs(String docs) {
+        if (docs == null) {
+            return "";
+        }
+
+        StringBuilder builder = new StringBuilder();
+        boolean lastEmpty = false;
+        boolean first = true;
+        for (String line : docs.replace("<p>", "\n").split("\n")) {
+            //process line by line
+            String trimmed = line.trim();
+            //if the lines are empty we only include a single empty line at most, and add a # character
+            if (trimmed.isEmpty()) {
+                if (!lastEmpty && !first) {
+                    lastEmpty = true;
+                    builder.append("\n");
+                }
+                continue;
+            }
+            //add the newlines
+            lastEmpty = false;
+            if (first) {
+                first = false;
+            } else {
+                builder.append("\n");
+            }
+            //replace some special characters, others are taken care of by regex below
+            builder.append(trimmed.replace("\n", "\n")
+                    .replace("<ul>", "")
+                    .replace("</ul>", "")
+                    .replace("<li>", " - ")
+                    .replace("</li>", ""));
+        }
+
+        String ret = builder.toString();
+        //replace @code
+        ret = Constants.JAVA_DOC_CODE_PATTERN.matcher(ret).replaceAll("`$1`");
+
+        //replace @see
+        ret = Constants.JAVA_DOC_SEE_PATTERN.matcher(ret).replaceAll("see `$1`");
+
+        //replace @link with a reference to the field name
+        Matcher matcher = Constants.JAVA_DOC_LINK_PATTERN.matcher(ret);
+        while (matcher.find()) {
+            ret = ret.replace(matcher.group(0), "`" + configify(matcher.group(1)) + "`");
+        }
+        return ret;
+    }
+
+    private String configify(String group) {
+        //replace uppercase characters with a - followed by lowercase
+        StringBuilder ret = new StringBuilder();
+        for (int i = 0; i < group.length(); ++i) {
+            char c = group.charAt(i);
+            if (Character.isUpperCase(c)) {
+                ret.append("-");
+                ret.append(Character.toLowerCase(c));
+            } else {
+                ret.append(c);
+            }
+        }
+        return ret.toString();
     }
 
     private static boolean hasParameterAnnotated(ExecutableElement ex, String annotationName) {
